Not generate example scripts if Skript folder already exists
Originally, when you delete the aliases, and the skript directory
is exists, the plugin will give errors.

We fixed that by adding a new condition,if skript directory is exists, but
it will create a bug: when we only delete the aliases (not scripts or anything),

it also creates the example scripts, which is annoying. This commit fixes that behaviour.

Also added some debug messages to find execute console command statements.

Signed-off-by: Mustafa ÖNCEL <support@lifemcserver.com>