Finally fix slow variable saving (resolves GH-82)
This speeds up variable saving a lot (like 8 minutes to 10 seconds for 720.000+ variables)
Because UUIDs are always known when OfflinePlayer/Player objects are created, but names must be read from the world data. It is stable as I tested it a lot, so I decided to make it default. Of course you can also disable it. See GH-82 for more information about this topic and problem.