More various stuff
Unfortunately I pushed the parser changes I'm working on
too early by accident, so the release is more delayed.

I planned it to be included in 2.2.19, 2.2.18 is planned
to be just stackoverflow fix, slow variable save fix and other changes,

not a - in diff - big parser change.

Here is some other various stuff, other bunch of stuff is also
waiting to be ready to push.